Reorg the folder structure for formatting and core service, fix naming space for core service after reorg
