+++ src/PoshKentico.Business/Configuration/Sites/GetCmsSiteCultureBusiness.cs
@@ -17,8 +17,8 @@
 
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
-using PoshKentico.Core.Services.Configuration;
-using PoshKentico.Core.Services.General;
+using PoshKentico.Core.Services.Configuration.Localization;
+using PoshKentico.Core.Services.Configuration.Sites;
 
 namespace PoshKentico.Business.Configuration.Sites
 {
