+++ src/PoshKentico.Business/Configuration/Sites/GetCmsSiteDomainAliasBusiness.cs
@@ -18,6 +18,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
 using PoshKentico.Core.Services.Configuration;
+using PoshKentico.Core.Services.Configuration.Sites;
 using PoshKentico.Core.Services.General;
 
 namespace PoshKentico.Business.Configuration.Sites
