+++ src/PoshKentico.Business/Configuration/Sites/SetCMSSiteBusiness.cs
@@ -18,7 +18,7 @@
 using System.ComponentModel.Composition;
 using CMS.SiteProvider;
 using ImpromptuInterface;
-using PoshKentico.Core.Services.Configuration;
+using PoshKentico.Core.Services.Configuration.Sites;
 
 namespace PoshKentico.Business.Configuration.Sites
 {
