+++ src/PoshKentico.Core/PoshKentico.Core.csproj
@@ -451,14 +451,14 @@
   <ItemGroup>
     <Compile Include="GlobalSuppressions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Providers\Configuration\KenticoSiteService.cs" />
+    <Compile Include="Providers\Configuration\Sites\KenticoSiteService.cs" />
     <Compile Include="Providers\General\KenticoCmsApplicationService.cs" />
     <Compile Include="Providers\Development\KenticoWebPartService.cs" />
     <Compile Include="Providers\General\KenticoSiteLocationCache.cs" />
-    <Compile Include="Services\Configuration\ICulture.cs" />
-    <Compile Include="Services\Configuration\ISite.cs" />
-    <Compile Include="Services\Configuration\ISiteDomainAlias.cs" />
-    <Compile Include="Services\Configuration\ISiteService.cs" />
+    <Compile Include="Services\Configuration\Localization\ICulture.cs" />
+    <Compile Include="Services\Configuration\Sites\ISite.cs" />
+    <Compile Include="Services\Configuration\Sites\ISiteDomainAlias.cs" />
+    <Compile Include="Services\Configuration\Sites\ISiteService.cs" />
     <Compile Include="Services\Development\IWebPartCategory.cs" />
     <Compile Include="Services\Development\IWebPartService.cs" />
     <Compile Include="Services\General\ICmsApplicationService.cs" />
