+++ src/PoshKentico.Tests/Configuration/Sites/GetCmsSiteBusinessTests.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 using FluentAssertions;
 using Moq;
 using NUnit.Framework;
-using PoshKentico.Core.Services.Configuration;
+using PoshKentico.Core.Services.Configuration.Sites;
 
 namespace PoshKentico.Business.Configuration.Sites.Tests
 {
