+++ src/PoshKentico.Tests/Configuration/Sites/RemoveCmsSiteBusinessTests.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 using Moq;
 using NUnit.Framework;
 using PoshKentico.Business.Configuration.Sites;
-using PoshKentico.Core.Services.Configuration;
+using PoshKentico.Core.Services.Configuration.Sites;
 
 namespace PoshKentico.Tests.Configuration.Sites
 {
