+++ src/PoshKentico.Tests/Configuration/Sites/SetCmsSiteBusinessTests.cs
@@ -20,8 +20,7 @@ using CMS.SiteProvider;
 using Moq;
 using NUnit.Framework;
 using PoshKentico.Business.Configuration.Sites;
-using PoshKentico.Core.Services.Configuration;
-using PoshKentico.Core.Services.General;
+using PoshKentico.Core.Services.Configuration.Sites;
 
 namespace PoshKentico.Tests.Configuration.Sites
 {
