+++ src/PoshKentico/FormattingFiles/Configuration/Sites/SiteDomainAliasInfo.format.ps1xml
@@ -1,0 +1,49 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<Configuration>
+  <ViewDefinitions>
+    <View>
+      <Name>Domain Alisa</Name>
+      <ViewSelectedBy>
+        <TypeName>CMS.SiteProvider.SiteDomainAliasInfo</TypeName>
+      </ViewSelectedBy>
+      <TableControl>
+        <TableHeaders>
+          <TableColumnHeader>
+            <Width>15</Width>
+            <Label>AliasID</Label>
+          </TableColumnHeader>
+          <TableColumnHeader>
+            <Width>25</Width>
+            <Label>AliasName</Label>
+          </TableColumnHeader>
+          <TableColumnHeader>
+            <Width>25</Width>
+            <Label>DefaultAliasPath</Label>
+          </TableColumnHeader>
+          <TableColumnHeader>
+            <Width>25</Width>
+            <Label>RedirectUrl</Label>
+          </TableColumnHeader>
+        </TableHeaders>
+        <TableRowEntries>
+          <TableRowEntry>
+            <TableColumnItems>
+              <TableColumnItem>
+                <PropertyName>SiteDomainAliasID</PropertyName>
+              </TableColumnItem>
+              <TableColumnItem>
+                <PropertyName>SiteDomainAliasName</PropertyName>
+              </TableColumnItem>
+              <TableColumnItem>
+                <PropertyName>SiteDomainDefaultAliasPath</PropertyName>
+              </TableColumnItem>
+              <TableColumnItem>
+                <PropertyName>SiteDomainRedirectUrl</PropertyName>
+              </TableColumnItem>
+            </TableColumnItems>
+          </TableRowEntry>
+        </TableRowEntries>
+      </TableControl>
+    </View>
+  </ViewDefinitions>
+</Configuration>
