--- src/PoshKentico/FormattingFiles/configuration.format.ps1xml
@@ -1,131 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<Configuration>
-  <ViewDefinitions>
-    <View>
-      <Name>Site</Name>
-      <ViewSelectedBy>
-        <TypeName>CMS.SiteProvider.SiteInfo</TypeName>
-      </ViewSelectedBy>
-      <TableControl>
-        <TableHeaders>
-          <TableColumnHeader>
-            <Width>15</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>10</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-          </TableColumnHeader>
-        </TableHeaders>
-        <TableRowEntries>
-          <TableRowEntry>
-            <TableColumnItems>
-              <TableColumnItem>
-                <PropertyName>SiteID</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>SiteName</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>Status</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>DomainName</PropertyName>
-              </TableColumnItem>
-            </TableColumnItems>
-          </TableRowEntry>
-        </TableRowEntries>
-      </TableControl>
-    </View>
-
-    <View>
-      <Name>Culture</Name>
-      <ViewSelectedBy>
-        <TypeName>CMS.Localization.CultureInfo</TypeName>
-      </ViewSelectedBy>
-      <TableControl>
-        <TableHeaders>
-          <TableColumnHeader>
-            <Width>15</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>15</Width>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-          </TableColumnHeader>
-        </TableHeaders>
-        <TableRowEntries>
-          <TableRowEntry>
-            <TableColumnItems>
-              <TableColumnItem>
-                <PropertyName>CultureID</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>CultureName</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>CultureCode</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>CultureShortName</PropertyName>
-              </TableColumnItem>
-            </TableColumnItems>
-          </TableRowEntry>
-        </TableRowEntries>
-      </TableControl>
-    </View>
-
-    <View>
-      <Name>Domain Alisa</Name>
-      <ViewSelectedBy>
-        <TypeName>CMS.SiteProvider.SiteDomainAliasInfo</TypeName>
-      </ViewSelectedBy>
-      <TableControl>
-        <TableHeaders>
-          <TableColumnHeader>
-            <Width>15</Width>
-            <Label>AliasID</Label>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-            <Label>AliasName</Label>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-            <Label>DefaultAliasPath</Label>
-          </TableColumnHeader>
-          <TableColumnHeader>
-            <Width>25</Width>
-            <Label>RedirectUrl</Label>
-          </TableColumnHeader>
-        </TableHeaders>
-        <TableRowEntries>
-          <TableRowEntry>
-            <TableColumnItems>
-              <TableColumnItem>
-                <PropertyName>SiteDomainAliasID</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>SiteDomainAliasName</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>SiteDomainDefaultAliasPath</PropertyName>
-              </TableColumnItem>
-              <TableColumnItem>
-                <PropertyName>SiteDomainRedirectUrl</PropertyName>
-              </TableColumnItem>
-            </TableColumnItems>
-          </TableRowEntry>
-        </TableRowEntries>
-      </TableControl>
-    </View>
-  </ViewDefinitions>
-</Configuration>
