+++ src/PoshKentico/PoshKentico.csproj
@@ -527,7 +527,13 @@
     <None Include="DSCResources\xWebPart\xWebPart.schema.mof">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <Content Include="FormattingFiles\configuration.format.ps1xml">
+    <Content Include="FormattingFiles\Configuration\Sites\SiteDomainAliasInfo.format.ps1xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="FormattingFiles\Configuration\Sites\SiteInfo.format.ps1xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="FormattingFiles\Configuration\Sites\CultureInfo.format.ps1xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <None Include="packages.config" />
