+++ src/PoshKentico/posh-kentico.psd1
@@ -62,7 +62,9 @@ CLRVersion = '4.0'
 # Format files (.ps1xml) to be loaded when importing this module
 FormatsToProcess = @( `
     "./FormattingFiles/development.format.ps1xml", `
-    "./FormattingFiles/configuration.format.ps1xml" `
+    "./FormattingFiles/Configuration/Sites/CultureInfo.format.ps1xml", `
+    "./FormattingFiles/Configuration/Sites/SiteDomainAliasInfo.format.ps1xml", `
+    "./FormattingFiles/Configuration/Sites/SiteInfo.format.ps1xml" `
 )
 
 # Modules to import as nested modules of the module specified in RootModule/ModuleToProcess
